PROCUCCIÓN 2022-06-03

+++ src/context/AppContext.tsx
@@ -7,7 +7,9 @@ import { check, PERMISSIONS, request, openSettings, PermissionStatus } from 'rea
 import { AppState as AplicationStatus, Platform } from 'react-native';
 import { useQuery } from 'react-query';
 import { validarJWT } from '../api/Api';
+import VersionNumber from 'react-native-version-number';
 import { getExpired, validateError } from '../functions/helpers';
+import { useVersionApp } from '../hooks/versionApp';
 const initialState: AppState = {
     versionApp: '1.1.0',
     status: 'no-loged',
@@ -18,6 +20,7 @@ const initialState: AppState = {
     expired: { hours: 99, minutes: 59, seconds: 59 },
     cameraPermissionStatus: 'unavailable',
     file: undefined,
+    isUpdate: false,
 }
 
 export const AppContext = createContext({} as AppContextProps);
@@ -25,26 +28,34 @@ export const AppContext = createContext({} as AppContextProps);
 export const AppProvider = ({ children }: any) => {
     const [state, dispatch] = useReducer(appReducer, initialState);
 
-    const JWT = useQuery(["JWT"], () => validarJWT(),
+    const versionApp = useVersionApp();
+    useQuery(["JWT"], () => validarJWT(),
         {
             retry: 0,
             onSuccess: async ({ Person, token, Service, AccountMW, directory }) => {
-                if (Person.id_role !== 1) {
-                    await AsyncStorage.clear();
-                    setMessage(validateError('No tienes acceso a este sistema'));
-                } else {
-                    await setPerson(Person, token, (directory) ? directory[0] : undefined);
-                    await setService(Service);
-                    if (Service) {
-                        const expired = getExpired(new Date(Service.exitDate));
-                        setExpired(expired);
-                    } else {
-                        setExpired(undefined);
-                    }
-                    if (AccountMW) {
-                        setAccount(AccountMW);
-                    }
-                }
+                await versionApp.refetch()
+                    .then(async data => {
+                        if (data.isError) {
+                            const meessage = validateError(`${data.error}`);
+                            if (meessage) setMessage({ message: meessage.message, type: 'error' });
+                        }
+                        if (data.isSuccess) {
+                            if (VersionNumber.appVersion === data.data.version) {
+                                if (Person.id_role !== 1) {
+                                    await AsyncStorage.clear();
+                                    setMessage(validateError('No tienes acceso a este sistema'));
+                                } else {
+                                    await setPerson(Person, token, (directory) ? directory[0] : undefined);
+                                    await setService(Service);
+                                    if (Service) {
+                                        const expired = getExpired(new Date(Service.exitDate));
+                                        setExpired(expired);
+                                    } else setExpired(undefined);
+                                    if (AccountMW) setAccount(AccountMW);
+                                }
+                            } else setUpdate(true);
+                        }
+                    }).catch(err => setMessage(validateError(`${err}`)))
             },
             onError: async error => {
                 if (state.message === undefined) {
@@ -56,15 +67,7 @@ export const AppProvider = ({ children }: any) => {
         }
     );
 
-    const validarToken = async () => {
-        const token = await AsyncStorage.getItem('token');
-        if (token !== null) {
-            JWT.refetch();
-        }
-    }
-
     useEffect(() => {
-        validarToken();
         AplicationStatus.addEventListener('change', state => {
             if (state !== 'active') return;
             checkCameraPermission();
@@ -100,6 +103,9 @@ export const AppProvider = ({ children }: any) => {
         else permissionStatus = await check(PERMISSIONS.ANDROID.CAMERA);
         dispatch({ type: 'setCameraPermission', payload: { cameraPermissionStatus: permissionStatus } })
     }
+    const setUpdate = async (isUpdate: boolean) => {
+        dispatch({ type: 'updateApp', payload: { isUpdate } });
+    }
 
     return (
         <AppContext.Provider
@@ -113,6 +119,7 @@ export const AppProvider = ({ children }: any) => {
                 logOut,
                 askCameraPermission,
                 checkCameraPermission,
+                setUpdate
             }}
         >
             {children}
